+++ src/main/java/com/example/gymcenter/services/TaiKhoanService.java
@@ -7,7 +7,7 @@ public interface TaiKhoanService {
     String addAdmin(Model model);
     String saveAdmin(TaiKhoan taiKhoan);
     String viewAdmin(Model model);
-    String updateAdmin(Integer id, Model model);
-    String deleteAdmin(Integer id);
+    String updateAdmin(Long id, Model model);
+    String deleteAdmin(Long id);
     String searchTaiKhoan(String term, Model model);
 }
